move behavior out of public API
