+++ pkg/controller/atlasproject/atlasproject_controller.go
@@ -18,6 +18,7 @@ package atlasproject
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 
 	"go.uber.org/zap"
@@ -332,4 +333,20 @@ func logIfWarning(ctx *workflow.Context, result workflow.Result) {
 	if result.IsWarning() {
 		ctx.Log.Warnw(result.GetMessage())
 	}
+}
+
+func lastSpecFrom(atlasProject *akov2.AtlasProject, annotation string) (*akov2.AtlasProjectSpec, error) {
+	var lastApplied akov2.AtlasProject
+	ann, ok := atlasProject.GetAnnotations()[annotation]
+
+	if !ok {
+		return nil, nil
+	}
+
+	err := json.Unmarshal([]byte(ann), &lastApplied.Spec)
+	if err != nil {
+		return nil, fmt.Errorf("error reading AtlasProject Spec from annotation [%s]: %w", annotation, err)
+	}
+
+	return &lastApplied.Spec, nil
 }
