+++ pkg/controller/atlasproject/atlasproject_controller_test.go
@@ -388,3 +388,36 @@ func TestFindProjectsForTeams(t *testing.T) {
 		})
 	}
 }
+
+func TestLastSpecFrom(t *testing.T) {
+	tests := map[string]struct {
+		annotations      map[string]string
+		expectedLastSpec *akov2.AtlasProjectSpec
+		expectedError    string
+	}{
+
+		"should return nil when there is no last spec": {},
+		"should return error when last spec annotation is wrong": {
+			annotations: map[string]string{"mongodb.com/last-applied-configuration": "{wrong}"},
+			expectedError: "error reading AtlasProject Spec from annotation [mongodb.com/last-applied-configuration]:" +
+				" invalid character 'w' looking for beginning of object key string",
+		},
+		"should return last spec": {
+			annotations: map[string]string{"mongodb.com/last-applied-configuration": "{\"name\": \"my-project\"}"},
+			expectedLastSpec: &akov2.AtlasProjectSpec{
+				Name: "my-project",
+			},
+		},
+	}
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			p := &akov2.AtlasProject{}
+			p.WithAnnotations(tt.annotations)
+			lastSpec, err := lastSpecFrom(p, "mongodb.com/last-applied-configuration")
+			if err != nil {
+				assert.ErrorContains(t, err, tt.expectedError)
+			}
+			assert.Equal(t, tt.expectedLastSpec, lastSpec)
+		})
+	}
+}
